Extract helm-releases plugin to separate project
Also made some internal types/functions public and moved them to "internal" sub-packages, because they are needed by other projects
