+++ helm-plugin/src/main/kotlin/org/unbrokendome/gradle/plugins/helm/command/HelmCommandSupport.kt
@@ -5,6 +5,9 @@ import org.gradle.api.provider.Provider
 import org.gradle.workers.WorkerExecutor
 import org.json.JSONArray
 import org.json.JSONObject
+import org.unbrokendome.gradle.plugins.helm.command.internal.GlobalHelmOptionsApplier
+import org.unbrokendome.gradle.plugins.helm.command.internal.HelmOptionsApplier
+import org.unbrokendome.gradle.plugins.helm.command.internal.HelmServerOptionsApplier
 import org.unbrokendome.gradle.plugins.helm.command.tasks.AbstractHelmCommandTask
 import org.unbrokendome.gradle.plugins.helm.model.Release
