+++ helm-plugin/src/main/kotlin/org/unbrokendome/gradle/plugins/helm/command/internal/HelmOptionsApplier.kt
@@ -1,0 +1,19 @@
+package org.unbrokendome.gradle.plugins.helm.command.internal
+
+import org.unbrokendome.gradle.plugins.helm.command.HelmExecSpec
+import org.unbrokendome.gradle.plugins.helm.command.HelmOptions
+
+
+/**
+ * Strategy interface to apply [HelmOptions] to a [HelmExecSpec].
+ */
+interface HelmOptionsApplier {
+
+    /**
+     * Applies a set of options to a [HelmExecSpec].
+     *
+     * @param spec the [HelmExecSpec]
+     * @param options the options to apply
+     */
+    fun apply(spec: HelmExecSpec, options: HelmOptions)
+}
