+++ helm-releases-plugin/src/main/kotlin/org/unbrokendome/gradle/plugins/helm/release/rules/HelmUninstallReleaseFromTargetTaskRule.kt
@@ -3,7 +3,7 @@ package org.unbrokendome.gradle.plugins.helm.release.rules
 import org.gradle.api.NamedDomainObjectCollection
 import org.gradle.api.tasks.TaskContainer
 import org.gradle.api.tasks.TaskDependency
-import org.unbrokendome.gradle.plugins.helm.command.setFrom
+import org.unbrokendome.gradle.plugins.helm.command.internal.setFrom
 import org.unbrokendome.gradle.plugins.helm.command.tasks.HelmUninstall
 import org.unbrokendome.gradle.plugins.helm.release.dsl.HelmRelease
 import org.unbrokendome.gradle.plugins.helm.release.dsl.HelmReleaseInternal
