+++ helm-releases-plugin/src/test/kotlin/org/unbrokendome/gradle/plugins/helm/release/spek/PropertyMappingTest.kt
@@ -1,4 +1,4 @@
-package org.unbrokendome.gradle.plugins.helm.spek
+package org.unbrokendome.gradle.plugins.helm.release.spek
 
 import assertk.assertThat
 import assertk.assertions.prop
